fixes #982
